Fixes for warnings in 10.14.  Most are related to deprecated methods or implied type conversion.

+++ Source/Experiments/Ncd/Detector/NcdTube.h
@@ -88,7 +88,7 @@ typedef  enum  tubeNameConstants { //don't change the order
 -(void) setObject:(id)obj forKey:(id)key;
 - (NSUndoManager*) undoManager;
 - (int) compareIntTo:(id)aTube usingKey:(NSString*)aKey;
-- (int) compareStringTo:(id)aTube usingKey:(NSString*)aKey;
+- (NSInteger) compareStringTo:(id)aTube usingKey:(NSString*)aKey;
 - (int) compareFloatTo:(id)aTube usingKey:(NSString*)aKey;
 - (int) sortTypeFor:(NSString*)aKey;
