+++ Source/Objects/Data Processing/Histogram Objects/Histogram2D/OR2DHistoController.m
@@ -113,7 +113,7 @@
 
 - (IBAction) hideShowControls:(id)sender
 {
-    unsigned int oldResizeMask = [containingView autoresizingMask];
+    NSUInteger oldResizeMask = [containingView autoresizingMask];
     [containingView setAutoresizingMask:NSViewMinYMargin];
 
     NSRect aFrame = [NSWindow contentRectForFrameRect:[[self window] frame]
