+++ Source/Objects/OrcaObject/OrcaObject.m
@@ -366,7 +366,7 @@ NSString* ORMiscAttributeKey		= @"ORMiscAttributeKey";
 }
 
 
-- (int) compareStringTo:(id)anElement usingKey:(NSString*)aKey
+- (NSInteger) compareStringTo:(id)anElement usingKey:(NSString*)aKey
 {
     NSString* ourKey   = [self valueForKey:aKey];
     NSString* theirKey = [anElement valueForKey:aKey];
